Extract overload hint formatting into public MethodSignatureFormatter
Move the overload-signature hint logic added in #128 (AppendOverloads,
FormatSignature, FormatType, SnakeCaseName, FormatDefaultValue) out of
MethodBinder into a new public MethodSignatureFormatter class so it can be
reused outside the binder. Downstream consumers (e.g. Lean) can now append
the same "The following overloads are available:" hint to their own
user-facing error messages when they reject a PyObject argument themselves.

- FormatOverloads(methods, maxShown = 10, displayName = null) returns the
  hint text ("The expected signature is:" / "The following overloads are
  available:" plus one signature per line), or an empty string. Still
  best-effort: it never throws.
- FormatSignature(method, displayName = null) is public; the optional
  displayName lets constructors render as the type name instead of the
  special .ctor token.
- MethodBinder behavior is unchanged: the "No method matches given
  arguments" message is byte-identical.
